+++ inst/examples/examples.R
@@ -2,12 +2,15 @@
 renderInlineCode(paste(examplePath(), "templates/template1.docx", sep = ""),
                  paste(examplePath(), "results/result1.docx", sep = ""))
 
-############################## adding Flex Tables example
-library(ReporteRs)
-ft_mtcars <- vanilla.table(mtcars)
-ft_iris <- vanilla.table(iris)
-FT <- list(ft_mtcars = ft_mtcars, ft_iris = ft_iris)
-addFlexTables(paste(examplePath(), "templates/templateFT.docx", sep = ""), paste(examplePath(), "results/resultFT.docx", sep = ""), FT)
+############################## adding flextables example
+library(flextable)
+ft_mtcars <- flextable(mtcars)
+ft_iris <- flextable(iris)
+FT <- list(ft_mtcars=ft_mtcars,ft_iris=ft_iris)
+body_add_flextables(
+  paste(examplePath(),'templates/templateFT.docx',sep = ''),
+  paste(examplePath(),'/resultFT.docx',sep = ''),
+  FT)
 
 ############################## adding Plots example
 library(ggplot2)
@@ -16,18 +19,18 @@ addPlots(paste(examplePath(), "templates/templatePlots.docx", sep = ""),
          paste(examplePath(), "results/resultPlots.docx", sep = ""), Plots, height = 4)
 
 ############################## combining all together
-library(ReporteRs)
+library(flextable)
 library(ggplot2)
 # prepare outputs
-ft_mtcars <- vanilla.table(mtcars)
-FT <- list(mtcars = ft_mtcars)
+ft_mtcars <- flextable(mtcars)
+FT <- list(mtcars=ft_mtcars)
 Plots <- list(mtcars1 = function() print(ggplot(mtcars, aes(x = wt, y = hp, col = as.factor(cyl))) + geom_point()))
 
 # render docx
 renderInlineCode(paste(examplePath(), "templates/templateAll.docx", sep = ""),
                  paste(examplePath(), "results/templateAll.docx", sep = ""))
 
-addFlexTables(paste(examplePath(), "results/templateAll.docx", sep = ""),
+body_add_flextables(paste(examplePath(), "results/templateAll.docx", sep = ""),
               paste(examplePath(), "results/templateAll.docx", sep = ""), FT)
 
 addPlots(paste(examplePath(), "results/templateAll.docx", sep = ""),