+++ NEWS.md
@@ -1,3 +1,9 @@
+# WordR 0.3
+* Removing dependancy on ReporteRs package (keeping it just in addFlexTables function for backward compatibility)
+* changing addPlots function (thanks D.Gohel) to use officer not ReporteRs
+* adding new function body_add_flextables() which should replace addFlexTables()
+* adding flextable into imports
+
 # WordR 0.2.3
 
 * changing the output path to tempdir() in examples and tests, as the CRAN tests were failing